sensor_fusion_queue: consistently use uint64_t for sensor types and ids

+++ corvis/src/tracker/sensor_fusion_queue.cpp
@@ -69,7 +69,7 @@ fusion_queue::~fusion_queue()
 
 std::string id_string(uint64_t global_id)
 {
-    int type = global_id / MAX_SENSORS, id = global_id % MAX_SENSORS;
+    uint64_t type = global_id / MAX_SENSORS, id = global_id % MAX_SENSORS;
     std::string type_string = "UNKNOWN";
     switch(type) {
         case rc_SENSOR_TYPE_IMAGE: type_string = "Camera"; break;
